feat(dashboard): feedback textarea, sidebar scroll fix, and code polish
- Extract AnnotateSidebar and AnnotateOverlay into annotateView.tsx
- Add feedback textarea (auto-grow, non-resizable) above Submit button
- Fix sidebar screenshot list scroll: thumbnails now flex-shrink:0
- CSS: softer textarea border, tighten feedback-to-submit spacing
- Address code review: eager imports, platform toBase64(), client-side
  tab data, remove dead formatting branch from runAnnotateClient
- Thread feedback through submit for both CLI (JSON) and zip paths
