+++ packages/dashboard/src/dashboard.tsx
@@ -16,13 +16,11 @@
 
 import React from 'react';
 import './dashboard.css';
-import { ChevronLeftIcon, ChevronRightIcon, DownloadIcon, LockIcon, LockOpenIcon, ReloadIcon, ScreenshotRegionIcon } from './icons';
-import { Annotations } from './annotations';
-import { buildAnnotatedImage, saveAnnotationAsDownload } from './annotationImage';
+import { ChevronLeftIcon, ChevronRightIcon, LockIcon, LockOpenIcon, ReloadIcon, ScreenshotRegionIcon } from './icons';
 import { clientToViewport, getImageLayout } from './imageLayout';
 import { Recording } from './recording';
+import { AnnotateSidebar, AnnotateOverlay } from './annotateView';
 
-import type { Annotation, AnnotationsHandle } from './annotations';
 import { ToolbarButton } from '@web/components/toolbarButton';
 import { useMeasureForRef } from '@web/uiUtils';
 
@@ -412,177 +410,5 @@ export const Dashboard: React.FC<DashboardProps> = ({ model }) => {
         />
       )}
     </main>
-  );
-};
-
-type AnnotateSidebarProps = {
-  model: DashboardModel;
-  session: NonNullable<DashboardModel['state']['annotateSession']>;
-  onSubmit: () => Promise<void> | void;
-  onClose: () => void;
-};
-
-const AnnotateSidebar: React.FC<AnnotateSidebarProps> = ({ model, session, onSubmit, onClose }) => {
-  const totalAnnotations = session.frames.reduce((n, f) => n + f.annotations.length, 0);
-  const [submitting, setSubmitting] = React.useState(false);
-
-  return (
-    <aside className='annotate-sidebar' aria-label='Annotation screenshots'>
-      <div className='annotate-sidebar-header dashboard-shell-sidebar-header'>
-        <h2 className='dashboard-shell-sidebar-title'>UI Review</h2>
-        <ToolbarButton
-          className='annotate-sidebar-close'
-          icon='close'
-          title='Close annotation'
-          onClick={onClose}
-        />
-      </div>
-      <div className='annotate-sidebar-list'>
-        {session.frames.map(frame => {
-          const selected = frame.id === session.selectedFrameId;
-          return (
-            <div
-              key={frame.id}
-              className={'annotate-sidebar-thumb' + (selected ? ' selected' : '')}
-            >
-              <button
-                className='annotate-sidebar-thumb-button'
-                onClick={() => model.toggleSelectFrame(frame.id)}
-                title={`${frame.sessionTitle || 'session'} · ${frame.tabTitle || 'tab'}\n${frame.url}`}
-                aria-pressed={selected}
-              >
-                <span
-                  className='annotate-sidebar-thumb-img-wrap'
-                  style={{ aspectRatio: `${frame.viewportWidth} / ${frame.viewportHeight}` }}
-                >
-                  <img
-                    className='annotate-sidebar-thumb-img'
-                    alt=''
-                    src={'data:image/png;base64,' + frame.data}
-                  />
-                  {frame.annotations.map(a => (
-                    <span
-                      key={a.id}
-                      className='annotate-sidebar-thumb-rect'
-                      style={{
-                        left: `${(a.x / frame.viewportWidth) * 100}%`,
-                        top: `${(a.y / frame.viewportHeight) * 100}%`,
-                        width: `${(a.width / frame.viewportWidth) * 100}%`,
-                        height: `${(a.height / frame.viewportHeight) * 100}%`,
-                        ['--annotation-color' as any]: a.color,
-                      }}
-                    />
-                  ))}
-                </span>
-              </button>
-              <ToolbarButton
-                className='annotate-sidebar-thumb-remove'
-                icon='close'
-                title='Remove screenshot'
-                onClick={e => {
-                  e.stopPropagation();
-                  model.removeAnnotateFrame(frame.id);
-                }}
-              />
-            </div>
-          );
-        })}
-      </div>
-      <button
-        className='annotate-sidebar-submit'
-        disabled={totalAnnotations === 0 || submitting}
-        onClick={async () => {
-          setSubmitting(true);
-          try {
-            await onSubmit();
-          } finally {
-            setSubmitting(false);
-          }
-        }}
-      >
-        {submitting ? 'Submitting…' : 'Submit'}
-      </button>
-    </aside>
-  );
-};
-
-type AnnotateOverlayProps = {
-  model: DashboardModel;
-  frame: import('./dashboardModel').AnnotateFrame;
-};
-
-const AnnotateOverlay: React.FC<AnnotateOverlayProps> = ({ model, frame }) => {
-  const annotationsRef = React.useRef<AnnotationsHandle>(null);
-  const displayRef = React.useRef<HTMLImageElement>(null);
-  const containerRef = React.useRef<HTMLDivElement>(null);
-
-  const onAnnotationsChange = React.useCallback((next: Annotation[]) => {
-    model.updateFrameAnnotations(frame.id, next);
-  }, [model, frame.id]);
-
-  const onSave = React.useCallback(async () => {
-    const img = displayRef.current;
-    if (!img)
-      return;
-    const blob = await buildAnnotatedImage(img, frame.viewportWidth, frame.viewportHeight, frame.annotations);
-    if (!blob)
-      return;
-    const safe = (frame.tabTitle || frame.url || 'screenshot').replace(/[^a-z0-9]+/gi, '-').slice(0, 40) || 'screenshot';
-    const stamp = new Date().toISOString().replace(/[:.]/g, '-');
-    await saveAnnotationAsDownload(blob, `annotations-${stamp}-${safe}.png`);
-  }, [frame]);
-
-  const onClear = React.useCallback(() => {
-    model.updateFrameAnnotations(frame.id, []);
-    annotationsRef.current?.clearSelection();
-  }, [model, frame.id]);
-
-  return (
-    <div className='annotate-overlay' role='dialog' aria-label={`Annotate screenshot from ${frame.tabTitle || frame.url || 'page'}`}>
-      <div className='annotate-overlay-window'>
-        <div className='annotate-overlay-chrome'>
-          <span className='annotate-overlay-titlebar'>
-            <span className='annotate-overlay-title-text'>{frame.tabTitle || 'untitled'}</span>
-            <span className='annotate-overlay-title-sep'>·</span>
-            <span className='annotate-overlay-title-url'>{frame.url}</span>
-          </span>
-          <ToolbarButton
-            title='Save screenshot'
-            onClick={onSave}
-          >
-            <DownloadIcon />
-          </ToolbarButton>
-          <ToolbarButton
-            title='Clear annotations'
-            icon='circle-slash'
-            disabled={frame.annotations.length === 0}
-            onClick={onClear}
-          />
-          <ToolbarButton
-            title='Close screenshot'
-            icon='close'
-            onClick={() => model.deselectFrame()}
-          />
-        </div>
-        <div ref={containerRef} className='annotate-overlay-canvas'>
-          <img
-            ref={displayRef}
-            className='annotate-modal-image'
-            alt='annotation'
-            src={'data:image/png;base64,' + frame.data}
-          />
-          <Annotations
-            ref={annotationsRef}
-            active={true}
-            displayRef={displayRef}
-            screenRef={containerRef}
-            viewportWidth={frame.viewportWidth}
-            viewportHeight={frame.viewportHeight}
-            annotations={frame.annotations}
-            onAnnotationsChange={onAnnotationsChange}
-          />
-        </div>
-      </div>
-    </div>
   );
 };
