+++ packages/playwright-core/src/tools/dashboard/dashboardController.ts
@@ -39,7 +39,7 @@ export class DashboardConnection implements Transport {
   private _attachedPage: AttachedPage | undefined;
   private _onclose: () => void;
   private _onconnected?: () => void;
-  private _onAnnotationSubmit?: (frames: SubmittedAnnotationFrame[]) => void;
+  private _onAnnotationSubmit?: (frames: SubmittedAnnotationFrame[], feedback: string) => void;
   private _pushTabsScheduled = false;
   private _visible = true;
   private _pendingReveal: { sessionName?: string; workspaceDir?: string; pageId?: string } | undefined;
@@ -48,7 +48,7 @@ export class DashboardConnection implements Transport {
   _recordingDir: string;
   _streams = new Map<string, { handle: fs.promises.FileHandle; path: string }>();
 
-  constructor(provider: SessionProvider, onclose: () => void, onconnected?: () => void, onAnnotationSubmit?: (frames: SubmittedAnnotationFrame[]) => void) {
+  constructor(provider: SessionProvider, onclose: () => void, onconnected?: () => void, onAnnotationSubmit?: (frames: SubmittedAnnotationFrame[], feedback: string) => void) {
     this._provider = provider;
     this._onclose = onclose;
     this._onconnected = onconnected;
@@ -163,8 +163,8 @@ export class DashboardConnection implements Transport {
     this._pushTabs();
   }
 
-  async submitAnnotation(params: { frames: SubmittedAnnotationFrame[] }) {
-    this._onAnnotationSubmit?.(params.frames);
+  async submitAnnotation(params: { frames: SubmittedAnnotationFrame[]; feedback: string }) {
+    this._onAnnotationSubmit?.(params.frames, params.feedback);
   }
 
   async reveal(params: { path: string }) {
@@ -433,12 +433,10 @@ class AttachedPage {
     return { streamId };
   }
 
-  async screenshot(): Promise<{ data: string; viewportWidth: number; viewportHeight: number; ariaSnapshot: string; sessionTitle: string; tabTitle: string; url: string }> {
+  async screenshot(): Promise<{ data: string; viewportWidth: number; viewportHeight: number; ariaSnapshot: string; sessionTitle: string }> {
     const buffer = await this._page.screenshot({ type: 'png' });
     const vp = await this._viewportSize();
     const ariaSnapshot = await this._page.ariaSnapshot({ boxes: true, mode: 'ai' });
-    const tabTitle = await this._page.title().catch(() => '');
-    const url = this._page.url();
     const sessionTitle = this._owner.sessionTitleFor(this._page.context());
     return {
       data: buffer.toString('base64'),
@@ -446,8 +444,6 @@ class AttachedPage {
       viewportHeight: vp.height,
       ariaSnapshot,
       sessionTitle,
-      tabTitle,
-      url,
     };
   }
